wallet/scan: move WalletRescanReserver to scan files
